refac(zk): extract `CircuitTest` from `SimplCircuitTest`
