+++ tachyon/zk/plonk/circuit/examples/BUILD.bazel
@@ -3,6 +3,13 @@ load("//bazel:tachyon_cc.bzl", "tachyon_cc_library", "tachyon_cc_unittest")
 package(default_visibility = ["//visibility:public"])
 
 tachyon_cc_library(
+    name = "circuit_test",
+    testonly = True,
+    hdrs = ["circuit_test.h"],
+    deps = ["//tachyon/zk/base/halo2:halo2_prover_test"],
+)
+
+tachyon_cc_library(
     name = "simple_circuit",
     hdrs = ["simple_circuit.h"],
     deps = [
@@ -17,8 +24,8 @@ tachyon_cc_unittest(
         "simple_circuit_unittest.cc",
     ],
     deps = [
+        ":circuit_test",
         ":simple_circuit",
-        "//tachyon/zk/base/halo2:halo2_prover_test",
         "//tachyon/zk/plonk/keys:proving_key",
         "//tachyon/zk/plonk/keys/halo2:pinned_verifying_key",
     ],
