+++ tachyon/zk/plonk/circuit/examples/simple_circuit_unittest.cc
@@ -2,8 +2,7 @@
 
 #include "gtest/gtest.h"
 
-#include "tachyon/math/elliptic_curves/bn/bn254/g1.h"
-#include "tachyon/zk/base/halo2/halo2_prover_test.h"
+#include "tachyon/zk/plonk/circuit/examples/circuit_test.h"
 #include "tachyon/zk/plonk/keys/halo2/pinned_verifying_key.h"
 #include "tachyon/zk/plonk/keys/proving_key.h"
 
@@ -11,46 +10,7 @@ namespace tachyon::zk {
 
 namespace {
 
-struct Point {
-  std::string_view x;
-  std::string_view y;
-};
-
-class SimpleCircuitTest : public Halo2ProverTest {
- protected:
-  static Commitment CreateCommitment(const Point& point) {
-    return Commitment(math::bn254::Fq::FromHexString(point.x),
-                      math::bn254::Fq::FromHexString(point.y));
-  }
-
-  static std::vector<Commitment> CreateCommitments(
-      const std::vector<Point>& points) {
-    return base::Map(points, &CreateCommitment);
-  }
-
-  static Evals CreateColumn(const std::vector<std::string_view>& column) {
-    std::vector<F> evaluations = base::Map(
-        column, [](std::string_view coeff) { return F::FromHexString(coeff); });
-    return Evals(std::move(evaluations));
-  }
-
-  static std::vector<Evals> CreateColumns(
-      const std::vector<std::vector<std::string_view>>& columns) {
-    return base::Map(columns, &CreateColumn);
-  }
-
-  static Poly CreatePoly(const std::vector<std::string_view>& poly) {
-    std::vector<F> coefficients = base::Map(
-        poly, [](std::string_view coeff) { return F::FromHexString(coeff); });
-    return Poly(math::UnivariateDenseCoefficients<F, kMaxDegree>(
-        std::move(coefficients)));
-  }
-
-  static std::vector<Poly> CreatePolys(
-      const std::vector<std::vector<std::string_view>>& polys) {
-    return base::Map(polys, &CreatePoly);
-  }
-};
+class SimpleCircuitTest : public CircuitTest {};
 
 }  // namespace
